more label cleanup
