+++ commit/metrics/prom.go
@@ -3,11 +3,10 @@ package metrics
 import (
 	"context"
 	"fmt"
-	"strconv"
 	"time"
 
 	"github.com/prometheus/client_golang/prometheus"
-	"github.com/prometheus/client_golang/prometheus/promauto"
+	chain_selectors "github.com/smartcontractkit/chain-selectors"
 	"github.com/smartcontractkit/chainlink-common/pkg/logger"
 	"go.opentelemetry.io/otel/attribute"
 	"go.opentelemetry.io/otel/metric"
@@ -22,87 +21,12 @@ import (
 	"github.com/smartcontractkit/chainlink-common/pkg/beholder"
 )
 
-var (
-	rmnLatencyBucketsMilliseconds = []float64{
-		5, 10, 25, 50, 100, 250, 500, 1000, 2500, 5000, 10000,
-	}
-	promProcessorOutputCounter = promauto.NewCounterVec(
-		prometheus.CounterOpts{
-			Name: "ccip_commit_processor_output_sizes",
-			Help: "This metric tracks the number of different items in the commit processor",
-		},
-		[]string{"chainFamily", "chainID", "processor", "method", "type"},
-	)
-	promProcessorLatencyHistogram = promauto.NewHistogramVec(
-		prometheus.HistogramOpts{
-			Name: "ccip_commit_processor_latency",
-			Help: "This metric tracks the client-observed latency of a single processor method",
-			Buckets: []float64{
-				float64(50 * time.Millisecond),
-				float64(100 * time.Millisecond),
-				float64(200 * time.Millisecond),
-				float64(500 * time.Millisecond),
-				float64(700 * time.Millisecond),
-				float64(time.Second),
-				float64(2 * time.Second),
-				float64(5 * time.Second),
-				float64(7 * time.Second),
-				float64(10 * time.Second),
-				float64(20 * time.Second),
-			},
-		},
-		[]string{"chainFamily", "chainID", "processor", "method"},
-	)
-	promProcessorErrors = promauto.NewCounterVec(
-		prometheus.CounterOpts{
-			Name: "ccip_commit_processor_errors",
-			Help: "This metric tracks the number of errors in the commit processor observation",
-		},
-		[]string{"chainFamily", "chainID", "processor", "method"},
-	)
-	promSequenceNumbers = promauto.NewGaugeVec(
-		prometheus.GaugeOpts{
-			Name: "ccip_commit_max_sequence_number",
-			Help: "This metric tracks the max sequence number observed by the commit processor",
-		},
-		[]string{"chainFamily", "chainID", "sourceChainFamily", "sourceChain",
-			"method", "source_network_name", "dest_network_name"},
-	)
-	promMerkleProcessorRmnReportLatency = promauto.NewHistogramVec(
-		prometheus.HistogramOpts{
-			Name:    "ccip_commit_merkle_processor_rmn_report_latency_ms",
-			Help:    "This metric tracks the client-observed latency of building an full RMN report with signatures",
-			Buckets: rmnLatencyBucketsMilliseconds,
-		},
-		[]string{"chainID", "success"},
-	)
-	promRmnControllerRmnRequestLatency = promauto.NewHistogramVec(
-		prometheus.HistogramOpts{
-			Name:    "ccip_commit_rmn_controller_rmn_request_latency_ms",
-			Help:    "This metric tracks the client-observed latency of a single RMN request",
-			Buckets: rmnLatencyBucketsMilliseconds,
-		},
-		[]string{"method", "nodeID", "error"},
-	)
-	promCommitLatestRoundID = promauto.NewGaugeVec(prometheus.GaugeOpts{
-		Name: "ccip_commit_latest_round_id",
-		Help: "The latest round ID observed by the commit plugin",
-	}, []string{"source_network_name", "dest_network_name", "contract_address", "plugin"})
-	promLooppCCIPProviderSupported = promauto.NewGaugeVec(prometheus.GaugeOpts{
-		Name: "ccip_commit_loopp_ccip_provider_supported",
-		Help: "Tracks whether LOOPP CCIP provider is supported for each chain family (1 = supported, 0 = not supported)",
-	}, []string{"chain_family"})
-	promCommitConfigDigestMismatch = promauto.NewGaugeVec(prometheus.GaugeOpts{
-		Name: "ccip_commit_config_digest_mismatch",
-		Help: "Reports whether the home chain config digest differs from the offramp config digest (1 = mismatch, 0 = match)",
-	}, []string{"chain_family", "chain_id"})
-)
-
 type PromReporter struct {
 	lggr        logger.Logger
 	bhClient    beholder.Client
 	chainFamily string
 	chainID     string
+
 	// Prometheus components
 	merkleProcessorRmnReportHistogram *prometheus.HistogramVec
 	rmnControllerRmnRequestHistogram  *prometheus.HistogramVec
@@ -113,7 +37,8 @@ type PromReporter struct {
 	commitLatestRound         *prometheus.GaugeVec
 	sequenceNumbers           *prometheus.GaugeVec
 	looppProviderSupported    *prometheus.GaugeVec
-	// Beholder components
+
+	// Beholder equivalents of the above.
 	bhProcessorLatencyHistogram metric.Int64Histogram
 	bhProcessorOutputCounter    metric.Int64Counter
 	bhProcessorErrors           metric.Int64Counter
@@ -321,30 +246,27 @@ func (p *PromReporter) TrackOutcome(outcome committypes.Outcome, round uint64) {
 func (p *PromReporter) trackLatestRoundID(
 	latestRoundID uint32, sourceChainSelector cciptypes.ChainSelector, onramp string, method string,
 ) {
-
+	// Could happen due to out of date chain-selectors lib.
 	sourceFamily, sourceChainID, ok := libs.GetChainInfoFromSelector(sourceChainSelector)
 	if !ok {
-		p.lggr.Errorw("failed to get chain ID from selector", "selector", sourceChainSelector)
-		return
+		// graceful fallback - we could even alert on such a thing.
+		sourceFamily = "unknown"
+		sourceChainID = "unknown"
 	}
 	sourceName, err := libs.GetNameFromIDAndFamily(sourceChainID, sourceFamily)
 	if err != nil {
-		p.lggr.Errorw("failed to get chain name from ID and family", "chainID",
-			sourceChainID, "family", sourceFamily, "err", err)
+		sourceName = "unknown"
 	}
 	destName, err := libs.GetNameFromIDAndFamily(p.chainID, p.chainFamily)
 	if err != nil {
-		p.lggr.Errorw("failed to get chain name from ID and family", "chainID",
-			p.chainID, "family", p.chainFamily, "err", err)
+		destName = "unknown"
 	}
 
 	p.commitLatestRound.WithLabelValues(sourceName, destName, onramp, method).Set(float64(latestRoundID))
-	p.bhCommitLatestRound.Record(context.Background(), int64(latestRoundID), metric.WithAttributes(
-		attribute.String("source_network_name", sourceName),
-		attribute.String("dest_network_name", destName),
-		attribute.String("contract_address", onramp),
-		attribute.String("plugin", method),
-	))
+
+	attrs := append(p.sourceChainAttrs(sourceChainSelector), p.destChainAttrs()...)
+	attrs = append(attrs, attribute.String("onrampAddress", onramp), attribute.String("method", method))
+	p.bhCommitLatestRound.Record(context.Background(), int64(latestRoundID), metric.WithAttributes(attrs...))
 }
 
 func (p *PromReporter) trackMaxSequenceNumber(
@@ -356,35 +278,29 @@ func (p *PromReporter) trackMaxSequenceNumber(
 		return
 	}
 
+	// Could happen due to out of date chain-selectors lib.
 	sourceFamily, sourceChainID, ok := libs.GetChainInfoFromSelector(sourceChainSelector)
 	if !ok {
-		p.lggr.Errorw("failed to get chain ID from selector", "selector", sourceChainSelector)
-		return
+		// graceful fallback - we could even alert on such a thing.
+		sourceFamily = "unknown"
+		sourceChainID = "unknown"
 	}
 	sourceName, err := libs.GetNameFromIDAndFamily(sourceChainID, sourceFamily)
 	if err != nil {
-		p.lggr.Errorw("failed to get chain name from ID and family", "chainID",
-			sourceChainID, "family", sourceFamily, "err", err)
+		sourceName = "unknown"
 	}
 	destName, err := libs.GetNameFromIDAndFamily(p.chainID, p.chainFamily)
 	if err != nil {
-		p.lggr.Errorw("failed to get chain name from ID and family", "chainID",
-			p.chainID, "family", p.chainFamily, "err", err)
+		destName = "unknown"
 	}
 
 	p.sequenceNumbers.
 		WithLabelValues(p.chainFamily, p.chainID, sourceFamily, sourceChainID, method, sourceName, destName).
 		Set(float64(maxSeqNr))
 
-	p.bhSequenceNumbers.Record(context.Background(), int64(maxSeqNr), metric.WithAttributes(
-		attribute.String("chainFamily", p.chainFamily),
-		attribute.String("chainID", p.chainID),
-		attribute.String("sourceChainFamily", sourceFamily),
-		attribute.String("sourceChainID", sourceChainID),
-		attribute.String("method", method),
-		attribute.String("source_network_name", sourceName),
-		attribute.String("dest_network_name", destName),
-	))
+	attrs := append(p.sourceChainAttrs(sourceChainSelector), p.destChainAttrs()...)
+	attrs = append(attrs, attribute.String("method", method))
+	p.bhSequenceNumbers.Record(context.Background(), int64(maxSeqNr), metric.WithAttributes(attrs...))
 
 	p.lggr.Debugw(
 		"exec latest max seq num",
@@ -395,16 +311,6 @@ func (p *PromReporter) trackMaxSequenceNumber(
 		"destChainFamily", p.chainFamily,
 		"maxSeqNr", maxSeqNr,
 	)
-}
-
-func (p *PromReporter) TrackRmnReport(latency float64, success bool) {
-	successStr := strconv.FormatBool(success)
-	p.merkleProcessorRmnReportHistogram.WithLabelValues(p.chainID, successStr).Observe(latency)
-}
-
-func (p *PromReporter) TrackRmnRequest(method string, latency float64, nodeID uint64, err string) {
-	nodeIDStr := strconv.FormatUint(nodeID, 10)
-	p.rmnControllerRmnRequestHistogram.WithLabelValues(method, nodeIDStr, err).Observe(latency)
 }
 
 func (p *PromReporter) TrackProcessorLatency(
@@ -423,12 +329,8 @@ func (p *PromReporter) TrackProcessorLatency(
 	p.processorLatencyHistogram.
 		WithLabelValues(p.chainFamily, p.chainID, processor, method).
 		Observe(float64(latency))
-	p.bhProcessorLatencyHistogram.Record(context.Background(), int64(latency), metric.WithAttributes(
-		attribute.String("chainFamily", p.chainFamily),
-		attribute.String("chainID", p.chainID),
-		attribute.String("processor", processor),
-		attribute.String("method", method),
-	))
+	attrs := append(p.destChainAttrs(), attribute.String("processor", processor), attribute.String("method", method))
+	p.bhProcessorLatencyHistogram.Record(context.Background(), int64(latency), metric.WithAttributes(attrs...))
 }
 
 func (p *PromReporter) TrackProcessorOutput(
@@ -440,13 +342,13 @@ func (p *PromReporter) TrackProcessorOutput(
 		p.processorOutputCounter.
 			WithLabelValues(p.chainFamily, p.chainID, processor, method, key).
 			Add(float64(val))
-		p.bhProcessorOutputCounter.Add(context.Background(), int64(val), metric.WithAttributes(
-			attribute.String("chainFamily", p.chainFamily),
-			attribute.String("chainID", p.chainID),
+		attrs := append(
+			p.destChainAttrs(),
 			attribute.String("processor", processor),
 			attribute.String("method", method),
 			attribute.String("type", key),
-		))
+		)
+		p.bhProcessorOutputCounter.Add(context.Background(), int64(val), metric.WithAttributes(attrs...))
 	}
 }
 
@@ -457,9 +359,8 @@ func (p *PromReporter) TrackLooppProviderSupported(looppCCIPProviderSupported ma
 			value = 1
 		}
 		p.looppProviderSupported.WithLabelValues(chainFamily).Set(value)
-		p.bhLooppProviderSupported.Record(context.Background(), int64(value), metric.WithAttributes(
-			attribute.String("chain_family", chainFamily),
-		))
+		attrs := append(p.destChainAttrs(), attribute.String("loopChainFamily", chainFamily))
+		p.bhLooppProviderSupported.Record(context.Background(), int64(value), metric.WithAttributes(attrs...))
 	}
 }
 
@@ -469,28 +370,42 @@ func (p *PromReporter) TrackConfigDigestMismatch(mismatch bool) {
 		value = 1
 	}
 	p.configDigestMismatch.WithLabelValues(p.chainFamily, p.chainID).Set(value)
-	p.bhConfigDigestMismatch.Record(context.Background(), int64(value), metric.WithAttributes(
-		attribute.String("chain_family", p.chainFamily),
-		attribute.String("chain_id", p.chainID),
-	))
+	p.bhConfigDigestMismatch.Record(
+		context.Background(),
+		int64(value),
+		metric.WithAttributes(p.destChainAttrs()...),
+	)
 }
 
 func (p *PromReporter) TrackPluginHeartbeat(phase string) {
-	p.bhPluginHeartbeat.Add(context.Background(), 1, metric.WithAttributes(
-		attribute.String("chainFamily", p.chainFamily),
-		attribute.String("chainID", p.chainID),
+	attrs := append(
+		p.destChainAttrs(),
 		attribute.String("phase", phase),
-	))
+	)
+	p.bhPluginHeartbeat.Add(context.Background(), 1, metric.WithAttributes(attrs...))
 }
 
 func (p *PromReporter) TrackReportValidationRejected(phase string, reason string) {
-	p.bhReportValidationRejected.Add(context.Background(), 1, metric.WithAttributes(
-		attribute.String("chainFamily", p.chainFamily),
-		attribute.String("chainID", p.chainID),
+	attrs := append(
+		p.destChainAttrs(),
 		attribute.String("phase", phase),
 		attribute.String("reason", reason),
-	))
+	)
+	p.bhReportValidationRejected.Add(context.Background(), 1, metric.WithAttributes(attrs...))
 }
+
+// Label key constants for attributes used in this reporter - to reduce the chance of drift.
+const (
+	destChainIDLabelKey       = "destChainID"
+	destChainFamilyLabelKey   = "destChainFamily"
+	destChainNameLabelKey     = "destChainName"
+	destChainSelectorLabelKey = "destChainSelector"
+
+	sourceChainIDLabelKey       = "sourceChainID"
+	sourceChainFamilyLabelKey   = "sourceChainFamily"
+	sourceChainNameLabelKey     = "sourceChainName"
+	sourceChainSelectorLabelKey = "sourceChainSelector"
+)
 
 // destChainAttrs returns this reporter's own (destination) chain-family/chain-ID attributes. The
 // commit plugin runs one instance per destination chain, so every per-lane metric needs this to stay
@@ -498,9 +413,21 @@ func (p *PromReporter) TrackReportValidationRejected(phase string, reason string
 // same Beholder/Prometheus pipeline -- without it, two plugin instances (e.g. dest=chain-B and
 // dest=chain-C) both reading source=chain-A emit identical label sets and collide into one series.
 func (p *PromReporter) destChainAttrs() []attribute.KeyValue {
+	// Could happen due to out of date chain-selectors lib.
+	destChainInfo, err := chain_selectors.GetChainDetailsByChainIDAndFamily(p.chainID, p.chainFamily)
+	if err != nil {
+		// graceful fallback - we could even alert on such a thing.
+		destChainInfo = chain_selectors.ChainDetails{
+			ChainSelector: 0,
+			ChainName:     "unknown",
+		}
+	}
+
 	return []attribute.KeyValue{
-		attribute.String("chainFamily", p.chainFamily),
-		attribute.String("chainID", p.chainID),
+		attribute.String(destChainIDLabelKey, p.chainID),
+		attribute.String(destChainFamilyLabelKey, p.chainFamily),
+		attribute.String(destChainNameLabelKey, destChainInfo.ChainName),
+		attribute.String(destChainSelectorLabelKey, fmt.Sprintf("%d", destChainInfo.ChainSelector)),
 	}
 }
 
@@ -509,22 +436,24 @@ func (p *PromReporter) destChainAttrs() []attribute.KeyValue {
 // consistently across metrics. Callers that identify a lane (source, but not dest) must also append
 // destChainAttrs() -- see its doc comment for why.
 func (p *PromReporter) sourceChainAttrs(sourceChainSelector cciptypes.ChainSelector) []attribute.KeyValue {
+	// Could happen due to out of date chain-selectors lib.
 	sourceFamily, sourceChainID, ok := libs.GetChainInfoFromSelector(sourceChainSelector)
 	if !ok {
-		p.lggr.Errorw("failed to get chain ID from selector", "selector", sourceChainSelector)
-		return []attribute.KeyValue{
-			attribute.String("sourceChainSelector", strconv.FormatUint(uint64(sourceChainSelector), 10)),
-		}
+		// graceful fallback - we could even alert on such a thing.
+		sourceFamily = "unknown"
+		sourceChainID = "unknown"
 	}
+
 	sourceName, err := libs.GetNameFromIDAndFamily(sourceChainID, sourceFamily)
 	if err != nil {
-		p.lggr.Errorw("failed to get chain name from ID and family", "chainID",
-			sourceChainID, "family", sourceFamily, "err", err)
+		sourceName = "unknown"
 	}
+
 	return []attribute.KeyValue{
-		attribute.String("sourceChainFamily", sourceFamily),
-		attribute.String("sourceChainID", sourceChainID),
-		attribute.String("source_network_name", sourceName),
+		attribute.String(sourceChainFamilyLabelKey, sourceFamily),
+		attribute.String(sourceChainIDLabelKey, sourceChainID),
+		attribute.String(sourceChainNameLabelKey, sourceName),
+		attribute.String(sourceChainSelectorLabelKey, fmt.Sprintf("%d", sourceChainSelector)),
 	}
 }
 
@@ -548,11 +477,8 @@ func (p *PromReporter) TrackRmnCurseActive(curseType string, active bool) {
 	if active {
 		value = 1
 	}
-	p.bhRmnCurseActive.Record(context.Background(), value, metric.WithAttributes(
-		attribute.String("chain_family", p.chainFamily),
-		attribute.String("chain_id", p.chainID),
-		attribute.String("curse_type", curseType),
-	))
+	attrs := append(p.destChainAttrs(), attribute.String("curse_type", curseType))
+	p.bhRmnCurseActive.Record(context.Background(), value, metric.WithAttributes(attrs...))
 }
 
 func (p *PromReporter) TrackSourceChainCursed(sourceChain cciptypes.ChainSelector, cursed bool) {
@@ -571,17 +497,11 @@ func (p *PromReporter) TrackObservationError(sourceChain cciptypes.ChainSelector
 }
 
 func (p *PromReporter) TrackFChainReadError() {
-	p.bhFChainReadErrors.Add(context.Background(), 1, metric.WithAttributes(
-		attribute.String("chainFamily", p.chainFamily),
-		attribute.String("chainID", p.chainID),
-	))
+	p.bhFChainReadErrors.Add(context.Background(), 1, metric.WithAttributes(p.destChainAttrs()...))
 }
 
 func (p *PromReporter) TrackConsensusObservationFailed() {
-	p.bhConsensusObservationFailed.Add(context.Background(), 1, metric.WithAttributes(
-		attribute.String("chain_family", p.chainFamily),
-		attribute.String("chain_id", p.chainID),
-	))
+	p.bhConsensusObservationFailed.Add(context.Background(), 1, metric.WithAttributes(p.destChainAttrs()...))
 }
 
 // TrackReportTransmissionGaveUp reports one source-chain lane within a report abandoned before
@@ -595,11 +515,8 @@ func (p *PromReporter) TrackReportTransmissionGaveUp(sourceChain cciptypes.Chain
 }
 
 func (p *PromReporter) TrackReportTransmissionAttempts(attempts uint, success bool) {
-	p.bhReportTransmissionAttempts.Record(context.Background(), int64(attempts), metric.WithAttributes(
-		attribute.String("chainFamily", p.chainFamily),
-		attribute.String("chainID", p.chainID),
-		attribute.Bool("success", success),
-	))
+	attrs := append(p.destChainAttrs(), attribute.Bool("success", success))
+	p.bhReportTransmissionAttempts.Record(context.Background(), int64(attempts), metric.WithAttributes(attrs...))
 }
 
 func (p *PromReporter) TrackRangeTruncated(sourceChain cciptypes.ChainSelector) {
@@ -631,13 +548,10 @@ func (p *PromReporter) TrackOffRampConsensusInsufficient(sourceChain cciptypes.C
 func (p *PromReporter) TrackConsensusDropped(
 	objectName string, key string, reason string, sourceChain cciptypes.ChainSelector,
 ) {
-	attrs := []attribute.KeyValue{
-		attribute.String("chainFamily", p.chainFamily),
-		attribute.String("chainID", p.chainID),
+	attrs := append(p.destChainAttrs(),
 		attribute.String("objectName", objectName),
 		attribute.String("key", key),
-		attribute.String("reason", reason),
-	}
+		attribute.String("reason", reason))
 	if sourceChain != 0 {
 		attrs = append(attrs, p.sourceChainAttrs(sourceChain)...)
 	}
